+++ samples/sqltransport/store-and-forward/Version_2/Receiver/Receiver.csproj
