+++ samples/sqltransport/store-and-forward/Version_2/SQLStoreAndForward.sln.DotSettings
