+++ samples/sqltransport/store-and-forward/Version_2/Sender/OrderAcceptedHandler.cs
