+++ samples/sqltransport/store-and-forward/Version_2/Shared/ConfigErrorQueue.cs
