+++ samples/sqltransport/store-and-forward/Version_3/Receiver/Receiver.csproj
