+++ samples/sqltransport/store-and-forward/Version_3/Receiver/app.config
@@ -11,13 +11,4 @@
       <add Assembly="Shared" Endpoint="Samples.SqlServer.StoreAndForwardSender" />
     </MessageEndpointMappings>
   </UnicastBusConfig>
-<startup><supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5.2" /></startup>
-  <runtime>
-    <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-      <dependentAssembly>
-        <assemblyIdentity name="NServiceBus.Core" publicKeyToken="9fc386479f8a226c" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-6.0.0.0" newVersion="6.0.0.0" />
-      </dependentAssembly>
-    </assemblyBinding>
-  </runtime>
 </configuration>
