+++ samples/sqltransport/store-and-forward/Version_3/Receiver/packages.config
