+++ samples/sqltransport/store-and-forward/Version_3/SQLStoreAndForward.sln.DotSettings
