+++ samples/sqltransport/store-and-forward/Version_3/Sender/ConnectionProvider.cs
