+++ samples/sqltransport/store-and-forward/Version_3/Sender/Sender.csproj
