+++ samples/sqltransport/store-and-forward/Version_3/Shared/Shared.csproj
@@ -49,7 +49,6 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ConfigErrorQueue.cs" />
     <Compile Include="OrderAccepted.cs" />
     <Compile Include="OrderSubmitted.cs" />
   </ItemGroup>
