+++ samples/sqltransport/store-and-forward/Version_3/Shared/packages.config
