--- samples/sqltransport/store-and-forward/Version_6/Shared/ConfigErrorQueue.cs
@@ -1,13 +1,0 @@
-using NServiceBus.Config;
-using NServiceBus.Config.ConfigurationSource;
-
-class ConfigErrorQueue : IProvideConfiguration<MessageForwardingInCaseOfFaultConfig>
-{
-    public MessageForwardingInCaseOfFaultConfig GetConfiguration()
-    {
-        return new MessageForwardingInCaseOfFaultConfig
-               {
-                   ErrorQueue = "error"
-               };
-    }
-}
